refactor(DamLabel): rename to DamItemLabel
