+++ packages/admin/cms-admin/src/dam/DataGrid/label/DamItemLabel.tsx
@@ -2,11 +2,11 @@ import { Typography } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import * as React from "react";
 
-import { MarkedMatches, TextMatch } from "../../common/MarkedMatches";
-import { GQLDamFileTableFragment, GQLDamFolderTableFragment } from "../../graphql.generated";
-import { isFile } from "../helpers/isFile";
-import { ArchivedTag } from "./tags/ArchivedTag";
-import { DamThumbnail } from "./thumbnail/DamThumbnail";
+import { MarkedMatches, TextMatch } from "../../../common/MarkedMatches";
+import { GQLDamFileTableFragment, GQLDamFolderTableFragment } from "../../../graphql.generated";
+import { isFile } from "../../helpers/isFile";
+import { ArchivedTag } from "../tags/ArchivedTag";
+import { DamThumbnail } from "../thumbnail/DamThumbnail";
 
 const LabelWrapper = styled("div")`
     display: flex;
@@ -34,7 +34,7 @@ const Path = styled(Typography)`
     color: ${({ theme }) => theme.palette.grey[300]};
 `;
 
-interface DamLabelProps {
+interface DamItemLabelProps {
     asset: GQLDamFolderTableFragment | GQLDamFileTableFragment;
     showPath?: boolean;
     matches?: TextMatch[];
@@ -59,7 +59,7 @@ const getFolderPath = (folder: GQLDamFolderTableFragment) => {
     return `/${pathArr.join("/")}`;
 };
 
-const DamLabel = ({ asset, showPath = false, matches }: DamLabelProps): React.ReactElement => {
+const DamItemLabel = ({ asset, showPath = false, matches }: DamItemLabelProps): React.ReactElement => {
     return (
         <LabelWrapper>
             <DamThumbnail asset={asset} />
@@ -76,4 +76,4 @@ const DamLabel = ({ asset, showPath = false, matches }: DamLabelProps): React.Re
     );
 };
 
-export default DamLabel;
+export default DamItemLabel;
